+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Doc guide generator
 - Add basic component library setup
 - Add injectable session URL rewrite function to support proxies
+- Add POSTLogger for meeting sessions
 
 ### Changed
 - Stop content share if the media stream end